wrap creation of compilerHosts - less code changes to master
